pyproject.toml
[tool.poetry]
name = "smart-spy-api"
version = "0.1.0"
description = ""
authors = ["Carlos Porta <cmaciasporta@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask-cors = "^3.0"
flask-praetorian = "^1.0"
fastapi = "^0.39.0"
uvicorn = "^0.9.0"
opencv-python = "^4.1"
python-multipart = "^0.0.5"
pyjwt = "^1.7"
passlib = {version = "^1.7", extras = ["bcrypt"]}
aiosqlite = "^0.10.0"
pony = "^0.7.10"
asgiref = "^3.2"
[tool.poetry.dev-dependencies]
pytest = "^3.0"
pylama = "^7.7"
pytest-asyncio = "^0.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
